Change calendar embedding.
The main reason for this is that Tockify's script forces the use of the
"mini" calendar widget on mobile, which doesn't provide instructions on
how to subscribe.

The other reason is because removing third-party Javascript code is
generally better for privacy, since it means that Tockify can't silently
add analytics code or such in the future.